pyproject.toml
[tool.poetry]
name = "oss-downloader"
version = "0.1.0"
description = ""
authors = ["webclerk <webclerk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
oss2 = "^2.18.3"
sqlalchemy = "^2.0.23"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
click = "^8.1.7"
uuid-utils = "^0.6.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
ruff = "^0.1.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
